vote proposal (#11)

+++ common/constants.go
@@ -48,7 +48,7 @@ const (
 	MaximumProposalExplainationLength = 1000
 	NumberOfDCBGovernors              = 50
 	NumberOfGOVGovernors              = 50
-	EncryptionPhaseDuration           = 5
+	EncryptionPhaseDuration           = uint32(5)
 )
 
 // board types
